Kill process when Tps_exe in BD is 1 or 0

+++ WindowsService1/WindowsService1/Service1.cs
@@ -8,14 +8,19 @@ using System.Linq;
 using System.ServiceProcess;
 using System.Text;
 using System.Threading.Tasks;
-using System.Collections.Generic;
+using System.Configuration;
+using System.Data;
+using System.Data.SqlClient;
 
 namespace WindowsService1
 {
     public partial class Service1 : ServiceBase
     {
         private int eventId = 1;
-
+        private string ConnectionString;
+        private List<String> appControlled;
+        private int counter;
+        private DataTable table;
 
         public Service1()
         {
@@ -37,10 +42,45 @@ namespace WindowsService1
 
         }
 
+        private void getApplicationControlled(string conn)
+        {
+            this.table = new DataTable();
+            this.appControlled = new List<string>();
+            
+            SqlConnection AppSmartConnection = new SqlConnection(conn);
+            AppSmartConnection.Open();
+            String request = "SELECT Nom_app FROM [dbo].[ApplicationControlable]";
+            // Création d'un objet commande basé sur la requête reçue en paramètre.
+            SqlCommand command = new SqlCommand(request, AppSmartConnection);
+            
+            SqlDataAdapter adapter = new SqlDataAdapter();
+            adapter.SelectCommand = command;
+
+            // Remplissage de la DataTable avec le résultat de la requête.
+            adapter.Fill(this.table);
+           
+            SqlDataReader reader = command.ExecuteReader();
+            if (reader.HasRows)
+            {
+                while (reader.Read())
+                {
+                    this.appControlled.Add(reader.GetString(0));
+                }
+            }
+            else
+            {
+                eventLog1.WriteEntry("No rows found.");
+            }
+            reader.Close();
+            AppSmartConnection.Close();
+        }
+
         protected override void OnStart(string[] args)
         {
             // Ecrire dans le journal des événements
             eventLog1.WriteEntry("Dans la méthode OnStart");
+            this.counter = 0;
+
             // On configure un timer pour se déclencher toutes les minutes.  
             System.Timers.Timer timer = new System.Timers.Timer();
             timer.Interval = 60000; // 60 secondes  
@@ -52,25 +92,73 @@ namespace WindowsService1
 
         public void OnTimer(object sender, System.Timers.ElapsedEventArgs args)
         {
+            eventLog1.WriteEntry("Dans la méthode onTimer");
+            if (this.counter<1)
+            {
+                LectureConfiguration();
+
+                getApplicationControlled(this.ConnectionString);
+                
+                this.counter += 1;
+                
+            }
+
             // Tracer le nombre d'exécution du traitement
             eventLog1.WriteEntry("Exécution du traitement", EventLogEntryType.Information, eventId++);
 
-            // TODO: Insérer ici le traitement que doit faire le service.
-
             Process[] processes = Process.GetProcesses();
-            Console.WriteLine("Count: {0}", processes.Length);
             WriteToFile("Nouveau Log");
+
+
+            List<String> processN = new List<string>();
             foreach(Process process in processes)
             {
-                WriteToFile(process.ProcessName);
-            }
-            List<String> toStop = readToFile();
-            foreach (String process in toStop)
-            {
-                foreach (Process p in Process.GetProcessesByName(process)) {
-                    p.Kill();
+                if (!processN.Contains(process.ProcessName))
+                {
+                    processN.Add(process.ProcessName);
+                    if (this.appControlled.Contains(process.ProcessName))
+                    {
+                        eventLog1.WriteEntry(process.ProcessName, EventLogEntryType.Information, eventId++);
+                        try
+                        {
+                            string requete = "SELECT Tps_exe_restant FROM ApplicationControlee ap" +
+                                " WHERE ap.Id_app = (SELECT Id_app FROM ApplicationControlable WHERE Nom_app = '" + process.ProcessName + "')";
+
+                            eventLog1.WriteEntry(requete, EventLogEntryType.Information, eventId++);
+                            // Initialiser la data source avec le résultat de la requête.
+                            int a = GetData(this.ConnectionString, requete);
+                            eventLog1.WriteEntry(a.ToString(), EventLogEntryType.Information, eventId++);
+
+                            if (a == 1)
+                            {
+                                foreach (Process p in Process.GetProcessesByName(process.ProcessName))
+                                {
+                                    p.Kill();
+                                }
+                            } else if (a > 1)
+                            {
+                                decreaseTpsExe(process.ProcessName);
+                            }
+
+
+                        }
+                        catch { }
+                    }
                 }
             }
+        }
+
+        private void decreaseTpsExe(string processName)
+        {
+            string requete = "UPDATE ApplicationControlee SET Tps_exe_restant = Tps_exe_restant - 1 WHERE Id_app = (SELECT id_app FROM ApplicationControlable WHERE Nom_app = '"+processName+"')";
+            // Création d'un objet connexion en se basant sur la chaine de connexion.
+            SqlConnection AppSmartConnection = new SqlConnection(this.ConnectionString);
+
+            // Création d'un objet commande basé sur la requête reçue en paramètre.
+            SqlCommand command = new SqlCommand(requete, AppSmartConnection);
+            AppSmartConnection.Open();
+            command.ExecuteNonQuery();
+            AppSmartConnection.Close();
         }
 
         public void WriteToFile(string Message)
@@ -97,6 +185,58 @@ namespace WindowsService1
                 }
             }
         }
+
+        private void LectureConfiguration()
+        {
+            // Lecture de la dernière chaine de connexion utilisée
+            ConnectionStringSettings oCnxCfg = ConfigurationManager.ConnectionStrings["DerniereChaine"];
+
+            if (oCnxCfg != null)
+            {
+                this.ConnectionString = oCnxCfg.ConnectionString;
+            }
+            else
+            {
+                // Recherche de la chaine de connexion partielle à la base de données dans le App.config
+                oCnxCfg = ConfigurationManager.ConnectionStrings["ConnexionAppSmartLocker"];
+
+                if (oCnxCfg != null)
+                {
+                    SqlConnectionStringBuilder oCnxBldr = new SqlConnectionStringBuilder(oCnxCfg.ConnectionString);
+
+                    if (string.IsNullOrEmpty(oCnxBldr.InitialCatalog))
+                    {
+                        oCnxBldr.InitialCatalog = "AppSmartLocker";
+                    }
+
+                    this.ConnectionString = oCnxBldr.ConnectionString;
+                }
+            }
+
+        }
+        private int GetData(string conn, string sqlCommand)
+        {
+            int q = -1;
+            // Création d'un objet connexion en se basant sur la chaine de connexion.
+            SqlConnection AppSmartConnection = new SqlConnection(conn);
+         
+
+            // Création d'un objet commande basé sur la requête reçue en paramètre.
+            SqlCommand command = new SqlCommand(sqlCommand, AppSmartConnection);
+            AppSmartConnection.Open();
+            try
+            {
+                q = (Int32)command.ExecuteScalar();
+            }
+            catch { }
+
+            AppSmartConnection.Close();
+            return q;
+
+        }
+
+
+
 
         public List<String> readToFile()
         {
@@ -131,7 +271,10 @@ namespace WindowsService1
             eventLog1.WriteEntry("Dans la méthode OnPause.");
         }
 
+        private void eventLog1_EntryWritten(object sender, EntryWrittenEventArgs e)
+        {
 
+        }
     }
 }
